refactor(*): Move tests to src places
