+++ src/lib/utils.spec.ts
@@ -3,7 +3,7 @@
 import * as assert from 'assert';
 import * as fs from 'fs';
 
-import * as utils from '../../lib/utils';
+import * as utils from './utils';
 
 describe('Utils', () => {
